dbeaver/pro#2706-reset-settings Moved reset settings to DBeaverActivator
