+++ plugins/org.jkiss.dbeaver.ui.statistics/src/org/jkiss/dbeaver/ui/statistics/UIStatisticsActivator.java
@@ -92,6 +92,7 @@ public class UIStatisticsActivator extends AbstractUIPlugin {
     public void start(BundleContext context) throws Exception {
         super.start(context);
         plugin = this;
+        preferences = new BundlePreferenceStore(getBundle());
     }
 
     @Override
@@ -109,9 +110,6 @@ public class UIStatisticsActivator extends AbstractUIPlugin {
     }
 
     public DBPPreferenceStore getPreferences() {
-        if (preferences == null) {
-            preferences = new BundlePreferenceStore(getBundle());
-        }
         return preferences;
     }
 }
